requirements.txt
gradio
openai
pandas
altair
python-dotenv

